fix: 파일경로명 resquest -> request 리펙토링 영향도
