+++ src/main/java/kr/mywork/domain/member/repository/MemberRepository.java
@@ -5,16 +5,16 @@ import java.util.Optional;
 import java.util.UUID;
 
 import kr.mywork.domain.member.model.Member;
-import kr.mywork.domain.member.service.dto.resquest.MemberCreateRequest;
+import kr.mywork.domain.member.service.dto.request.MemberCreateRequest;
 
-public interface MemberRepository{
-    Optional<Member> findByEmailAndDeletedFalse(String email);
+public interface MemberRepository {
+	Optional<Member> findByEmailAndDeletedFalse(String email);
 
-	List<Member> findMemberByCompanyId(UUID companyId, int page,int memberPageSize);
+	List<Member> findMemberByCompanyId(UUID companyId, int page, int memberPageSize);
 
 	long countByCompanyIdAndDeletedFalse(UUID companyId);
 
-    Member save(MemberCreateRequest memberCreateRequest);
+	Member save(MemberCreateRequest memberCreateRequest);
 
 	boolean existsByEmail(String email);
