+++ src/main/java/kr/mywork/interfaces/member/controller/dto/request/MemberCreateWebRequest.java
@@ -1,0 +1,27 @@
+package kr.mywork.interfaces.member.controller.dto.request;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+import java.util.UUID;
+
+import kr.mywork.domain.member.service.dto.request.MemberCreateRequest;
+
+public record MemberCreateWebRequest(
+	UUID id,
+	UUID companyId,
+	String name,
+	String department,
+	String position,
+	String role,
+	String phoneNumber,
+	String email,
+	LocalDateTime birthDate
+) {
+	public MemberCreateRequest toServiceDto() {
+		String password = birthDate.format(DateTimeFormatter.ofPattern("yyyyMMdd"));
+		return new MemberCreateRequest(
+			id, companyId, name, department, position, role,
+			phoneNumber, email, birthDate, password
+		);
+	}
+}
